refactor paths.AgentVaultPath(): distinguish between file path and entry name
